% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCBeta.R
\name{beta_cdf}
\alias{beta_cdf}
\title{A function that generates random Numbers using the monte carlo method}
\usage{
beta_cdf(n, a, b)
}
\arguments{
\item{n}{The number of random Numbers generated}

\item{a}{The lower bound of the interval}

\item{b}{The upper bound of the interval}
}
\value{
a random sample of size \code{n}
}
\description{
a function to compute a Monte Carlo estimate of the Beta(3, 3) cdf.
              and use the function to estimate F(x) for x = 0.1,0.2,…,0.9. Compare the estimates with the values returned by the pbeta function in R.
}
\examples{
\dontrun{
a <- data.frame(x=seq(0.1,0.9,0.1),Montacarlo=numeric(9),pbeta=numeric(9))#construct a dataframe to show the value of cdf
i <- 1
while (i<=9) {
  a[i,2] <- beta_cdf(10000,0,i*0.1)
  a[i,3] <- pbeta(i*0.1,3,3)
  i=i+1
}
print(a)
}
}
