% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_reBeta.R
\name{betafunction}
\alias{betafunction}
\title{The function that generates random Numbers with the accept-rejection method}
\usage{
betafunction(n, a, b)
}
\arguments{
\item{n}{The number of random Numbers generated}

\item{a}{Parameter of beta distribution}

\item{b}{Parameter of beta distribution}
}
\value{
a random sample of size \code{n}
}
\description{
a function to generate a random sample of size n from the Beta(a,b) distribution by the acceptance-rejection method.
}
\examples{
\dontrun{
sample_beta <- betafunction(1e3,3,2)#record the sample
head(sample_beta,20)
hist(sample_beta)
}
}
